renamed regex.py into string.py
