+++ bokeh/models/map_plots.py
@@ -21,7 +21,16 @@ log = logging.getLogger(__name__)
 # Bokeh imports
 from ..core.enums import MapType
 from ..core.has_props import abstract
-from ..core.properties import JSON, Base64String, Bool, Enum, Float, Instance, Int, Override
+from ..core.properties import (
+    JSON,
+    Base64String,
+    Bool,
+    Enum,
+    Float,
+    Instance,
+    Int,
+    Override,
+)
 from ..core.validation import error, warning
 from ..core.validation.errors import (
     INCOMPATIBLE_MAP_RANGE_TYPE,
